Refactored tests to create common abstract class and made OpenTelemetry field private in the appender
